requirements.txt
sqlalchemy
psycopg2-binary
python-dotenv
pandas
pymatgen
mp-api
ase
torch
numpy
